ai_service/requirements.txt
# RepoTalk AI service dependencies
# Install from C:\dev\RepoTalk\ai_service with:
#   pip install -r requirements.txt

# Core API
fastapi>=0.115.3
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# LLM + embeddings
openai==1.50.0
sentence-transformers==3.1.0
torch==2.6.0
einops==0.8.0
tiktoken==0.7.0

# Vector store
qdrant-client==1.11.0

# Repo parsing
tree-sitter==0.23.0
tree-sitter-python==0.23.0
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
tree-sitter-java==0.23.0
tree-sitter-go==0.23.0

# Git cloning
GitPython==3.1.43

# Test / local verification utilities
httpx==0.27.0
pytest==8.3.0
pytest-cov==5.0.0

# Evaluation
ragas==0.1.20
langchain-groq==0.1.9
langchain-community==0.2.16
datasets==2.21.0
